Stop using truecypt safe for password file
I'm now using an asymmetrically encrypted org file with emacs. I
would be using a symmetrically encrypted file like I used to with
vim, but gpg-agent apparently can't cache the password. It seems that
I don't have to enter the password again when decrypting, but when
saving, I'm always asked to enter a password twice to re-encrypt.
Setting epa-file-cache-passphrase-for-symmetric-encryption is broken
(maybe it works with a patch), and the documentation says that using
gpg-agent is safer... I don't mind since I'd really rather not lose
my private key regardless, but using vim's encryption is probably
the more straightforward method.