Emacs - Add no-littering package
Also use use-package for savehist and don't load it during init (it doesn't need
to be called until after init; on the first init when packages are being cloned,
the timer will be called during initialization otherwise, which is unnecessary).
In general, I should move towards using more specific hooks for enabling some of
these modes.

Previously, the directory savehist-file was in was not created by default (which
would cause an error with a clean ~/.emacs.d).