Optimus-manager - Switch to optimus manager again (needs testing)
nvidia-xrun has had no updates since 2019 and recently had issues that required
changes to nvidia-xorg.conf for X to start:
https://github.com/Witko/nvidia-xrun/issues/160

It also had started to crash more often (especially when sleeping or resuming
from sleep with a monitor connected). I switched back to nivida-xrun because
optimus-manager was randomly crashing (rather than predictaly crashing). So far,
optimus-manager has not crashed for me. It is actively maintained, and the
hybrid mode makes it worth trying to switch.

- Use hybrid mode (potentially more battery life)
  - Having to explicitly choose which programs use the GPU is a bit of a pain,
    but mpv and steam should be fine for now
  - Run mpv with prime-run
  - Maybe run picom with prime-run (not for now)
- Add gpu-switch script (to remind to run sudo prime-switch)
- Run optimus manager setup in .xinitrc if not using nvidia-xrun
- Don't hardcode monitor name in hdmiadd and related zsh functions
- Add old and new autorandr configuration