Nsteam - Blacklist hidraw
Linux now has switch controller support builtin.  Steam's controller support
doesn't work correctly (it works but it acts as a keyboard not a controller).
Using the builtin support and disabling switch controller support in steam
allows using the controller as a controller in game (even through proton).  It
does prevent using the controller in steam though.